Bugfix: Use correct copyright date.
I noticed that I used the wrong year for the first copyright reference by accident.
I reviewed the original commit in the Sourceforge repository and the date is 2007/03/07.
That means the copyright is 2007 rather than 2011.
This is when the FLL project and its specifications were drafted and first committed to a source code repository.