Update: Separate contexts between news and index page and add additional message to access denied page.
Have the links go to news.html rather than the site index.
If for some reason the news page and the site index ever change, then these links do not need to be changed.

Some links are intended to got to the site index, such as the access denied and not found pages.
Most other links are intended to go to the news page.