Update: Add Design and Principles pages.
This prepares the website for the additional documentation on the FLL project.
Many of the text files from the "documents" directory will be transcribed into pages under this.
One such example would be the style_guide.txt file.

Includes miscellaneous changes.

Update the sources, preparing another website download package as of this commit.