Update: Fixes, improvements, and add additional specifications and documentation.
A number of website mistakes are fixed.
Tweak the styling some.
Implement a "highlight" class rather than hard-coding the highlighting of the menu entries against "unlink".

Add numerous specification and documentation pages.
This is a manual process because I have yet to implement an auto-generator script for this.
I have not determined the style practice as well.
I also reached some burn-out while writing this and started working faster and sloppier.
The result is a rather inconsistent mix of the HTML markup transcription of the specifications and documentation.
Due to being a bit sloppier near the end, I also expect several mistakes in the transcription process.

I did not get to the Featureless Make specifications and documentation within the scope of this commit.