Bugfix: The FLL page needs to reference the stable release and fix the change log.
The 0.6.0 is a stable release.
Fix the text of the fll page to reference this.

The "Add support for ..." change log entry should not be under Security Fixes and instead should be under Features.

The current date is still the 31st in my time zone, so I am going to replace the existing release tar ball with these changes.