Update: Development update of TKIS, tools, and development iso on 2023/05/09.
A collection of the latest state of the Turtle Kevux development.
This is not intended to work per-say but is instead a bookmark of the current progress.

Major work is needed in regards to the kernel, the initrd, and the system init.
There is no GUI work at all (such as Xorg, etc..).