Update: Specifications and add FSS-0010 Encrypted Simple Packet specification.
Add "version date" paragraph to each specification to make it clearer when this specification was last updated.

I happened to notice that there is a mistake in the FSS-000E specification.
I fixed this in the proper FLL source code but this is after the formal 0.6.8 release was made, unfortunately.
This documentation will be more up to date that the current release.

Update the FSS-000E and FSS-000F specifications to reflect the changes formally introduced as of the 0.6.8 release on 2023/12/31.

The FSS-0010 Encrypted Simple Packet specification has been present for a while now but I failed to update the website.
This specification is now present.