Update: GPG key checksum and gpg files.
I forgot to update these with the previous related commit: 48b518e6390cb54b251a626c7d473e8724be6bb6.

The deprecated key is still not quite expired (it expires next month), so use that one to sign this one again.
Future versions will not do this and will very likely be self-signed (there is not much that can be done about that).